✨ feat: Reorganaizing app.py, changed to using blueprints
