+++ app/blueprints/__init__.py
@@ -1,0 +1,3 @@
+"""
+Blueprint package for organizing app routes
+"""
