+++ app/templates/components/_navbar.html
@@ -68,7 +68,7 @@
 
             <!-- Login/Logout Button -->
             {% if current_user.is_authenticated %}
-                <a href="{{ url_for('logout') }}" class="btn btn-outline-primary ms-2">Logout</a>
+                <a href="{{ url_for('auth.logout') }}" class="btn btn-outline-primary ms-2">Logout</a>
             {% else %}
                 <!-- Trigger/Login Modal Button -->
                 <button type="button" class="btn btn-primary ms-3" data-bs-toggle="modal" data-bs-target="#loginModal">
